requirements.txt
faiss-cpu==1.9.0.post1
duckduckgo-search==7.3.0
huggingface-hub==0.27.1
langchain==0.3.17
langchain-community==0.3.15
langchain-core==0.3.33
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
python-dotenv==1.0.1
sentence-transformers==3.4.0
sentencepiece==0.2.0
streamlit==1.42.0
trafilatura==2.0.0
